Run the controller inside a kind cluster
These tests run against a kind cluster. In order for the controller to
run, the kind package requires the docker binary and a active docker
socket. To achieve this we have a static docker binary in the image and
mount the docker socket in the deployment using a hostPath. This means
that the kind cluster must have the docker socket from the host
bind-mounted into it's nodes. The example config in `tests/assets`
assumes only 1 node, but if more are needed - all of them should mount
it, or the deployment needs to have a node selector.
Additionally having the docker socket mounted as a hostPath means that
the container running the controller must be root.